refactor: 移动【玄剑】到新的独立卡牌包
