+++ lang/zh_CN/Package/ExpansionCardPackage.lua
@@ -27,6 +27,4 @@ return {
     ['city_under_siege'] = '兵临城下',
     [':city_under_siege'] = '锦囊牌<br/><b>时机：</b>出牌阶段<br/><b>目标：</b>一名其他角色<br/><b>\
     效果：</b>你展示牌堆顶的四张牌，依次对目标角色使用其中的【杀】，然后将其余的牌以原顺序放回牌堆顶',
-    ['xuanjian_sword'] = '玄剑',
-    [':xuanjian_sword'] = '装备牌·武器<br/><b>攻击范围</b>：3<br/><b>武器技能</b>：你可以将一种花色的所有手牌当作普通【杀】使用',
 }
