+++ lang/zh_CN/Package/OldFriendCardPackage.lua
@@ -1,0 +1,7 @@
+-- translation for OldFriend Card Package
+
+return {
+    ['OldFriendCardPackage'] = '老友季卡牌包',
+    ['xuanjian_sword'] = '玄剑',
+    [':xuanjian_sword'] = '装备牌·武器<br/><b>攻击范围</b>：3<br/><b>武器技能</b>：你可以将一种花色的所有手牌当作普通【杀】使用',
+}
